requirements.txt
numpy
matplotlib
scipy
Pillow
pyttsx3
networkx
speechrecognition
requests
openai